Update : Socket logic
