+++ pages/_app.tsx
@@ -13,8 +13,8 @@ import { useRouter } from 'next/router';
 import Script from 'next/script';
 import { AlarmBar } from '@/components/common/AlarmBar';
 import { Provider } from 'react-redux';
+import SocketManager from '@/components/common/Socket';
 const queryClient = new QueryClient();
-const ENDPOINT = `${process.env.NEXT_PUBLIC_LOCAL_SERVER}`;
 
 function MyApp({ Component, pageProps }: AppProps) {
   const [notification, setNotification] = useState<string | null>(null);
@@ -29,54 +29,15 @@ function MyApp({ Component, pageProps }: AppProps) {
   ];
   const isExcludedPage = excludedPages.includes(router.pathname);
 
-  useEffect(() => {
-    // 로그인 성공 후 Socket.IO 클라이언트를 생성하고 서버에 연결합니다.
-    const socket = socketIOClient(ENDPOINT);
-
-    // 서버로부터 "notify-post" 이벤트를 수신하면 알림을 설정합니다.
-    socket.on('notify-post', (data) => {
-      setNotification(data.message);
-    });
-
-    socket.on('loginSuccess', (data) => {
-      const { email, nickname } = data;
-      setNotification(
-        `User ${nickname} with email ${email} logged in successfully.`
-      );
-
-      setTimeout(() => {
-        setNotification(null);
-      }, 5000);
-    });
-
-    socket.on('random-posts', (posts) => {
-      setRandomPosts(posts);
-
-      // 2초 후에 게시물 숨기기
-      setTimeout(() => {
-        setRandomPosts(null);
-      }, 10000);
-    });
-
-    socket.on('newPost', (latestPosts) => {
-      setNotification(`새 글이 등록 되었습니다.`);
-    });
-
-    // socket.on('latest-posts', (latestPosts) => {
-
-    // })
-
-    // 컴포넌트가 언마운트될 때 소켓 연결을 닫습니다.
-    return () => {
-      socket.disconnect();
-    };
-  }, []);
-
   return (
     <>
       <Provider store={store}>
         {process.env.NEXT_PUBLIC_GOOGLE_ANALYTICS && (
           <>
+            <SocketManager
+              setNotification={setNotification}
+              setRandomPosts={setRandomPosts}
+            />
             <Script
               strategy="afterInteractive"
               src={`https://www.googletagmanager.com/gtag/js?id=${process.env.NEXT_PUBLIC_GOOGLE_ANALYTICS}`}
